+++ macos/Flutter/ephemeral/flutter_export_environment.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/opt/homebrew/Caskroom/flutter/3.27.1/flutter"
+export "FLUTTER_ROOT=/opt/homebrew/share/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/gabrielriccimazuqueli/Documents/Projetos/Ricci/Flutter/selectpicker"
 export "COCOAPODS_PARALLEL_CODE_SIGN=true"
 export "FLUTTER_BUILD_DIR=build"